backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.4
supabase==1.1.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pyjwt==2.8.0